requirements.txt
argparse==1.2.1
bottle==0.11.6
h5py==2.2.0b1
ipdb==0.7
ipython==0.13.1
numpy==1.7.0
pandas==0.10.1
pymongo==2.5.1
python-dateutil==2.1
pytz==2012j
requests==1.2.0
scipy==0.11.0
six==1.2.0
wsgiref==0.1.2
